test(agent-bff): fold the fetch restore into the stub helper
